requirements.txt
azure-functions==1.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
dateutils==0.6.12
idna==3.4
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
urllib3==1.26.14
jugaad-data==0.24
